requirements.txt
aiohttp
aiofiles
tqdm
python-dotenv
colorama
